Split dense conditionals into separate module
